+++ logs_sdk/lib/opentelemetry/sdk/logs.rb
@@ -7,8 +7,11 @@
 require_relative 'logs/version'
 require_relative 'logs/logger'
 require_relative 'logs/logger_provider'
+require_relative 'logs/log_record'
+require_relative 'logs/log_record_data'
 require_relative 'logs/log_record_processor'
 require_relative 'logs/export'
+require_relative 'logs/log_record_limits'
 
 module OpenTelemetry
   module SDK