+++ logs_sdk/lib/opentelemetry/sdk/logs/logger_provider.rb
@@ -9,8 +9,6 @@ module OpenTelemetry
     module Logs
       # The SDK implementation of OpenTelemetry::Logs::LoggerProvider.
       class LoggerProvider < OpenTelemetry::Logs::LoggerProvider
-        attr_reader :resource
-
         UNEXPECTED_ERROR_MESSAGE = 'unexpected error in ' \
           'OpenTelemetry::SDK::Logs::LoggerProvider#%s'
 
@@ -20,10 +18,13 @@ module OpenTelemetry
         #
         # @param [optional Resource] resource The resource to associate with
         #   new LogRecords created by {Logger}s created by this LoggerProvider.
+        # @param [optional LogRecordLimits] log_record_limits The limits for
+        #   attributes count and attribute length for LogRecords.
         #
         # @return [OpenTelemetry::SDK::Logs::LoggerProvider]
-        def initialize(resource: OpenTelemetry::SDK::Resources::Resource.create)
+        def initialize(resource: OpenTelemetry::SDK::Resources::Resource.create, log_record_limits: LogRecordLimits::DEFAULT)
           @log_record_processors = []
+          @log_record_limits = log_record_limits
           @mutex = Mutex.new
           @resource = resource
           @stopped = false
@@ -119,6 +120,35 @@ module OpenTelemetry
 
             results.max || Export::SUCCESS
           end
+        end
+
+        # @api private
+        def on_emit(timestamp: nil,
+                    observed_timestamp: nil,
+                    severity_text: nil,
+                    severity_number: nil,
+                    body: nil,
+                    attributes: nil,
+                    trace_id: nil,
+                    span_id: nil,
+                    trace_flags: nil,
+                    instrumentation_scope: nil,
+                    context: nil)
+
+          log_record = LogRecord.new(timestamp: timestamp,
+                                     observed_timestamp: observed_timestamp,
+                                     severity_text: severity_text,
+                                     severity_number: severity_number,
+                                     body: body,
+                                     attributes: attributes,
+                                     trace_id: trace_id,
+                                     span_id: span_id,
+                                     trace_flags: trace_flags,
+                                     resource: @resource,
+                                     instrumentation_scope: instrumentation_scope,
+                                     log_record_limits: @log_record_limits)
+
+          @log_record_processors.each { |processor| processor.on_emit(log_record, context) }
         end
       end
     end