+++ logs_sdk/test/opentelemetry/sdk/logs/logger_provider_test.rb
@@ -19,7 +19,16 @@ describe OpenTelemetry::SDK::Logs::LoggerProvider do
 
     it 'allows a resource to be associated with the logger provider' do
       assert_instance_of(
-        OpenTelemetry::SDK::Resources::Resource, logger_provider.resource
+        OpenTelemetry::SDK::Resources::Resource, logger_provider.instance_variable_get(:@resource)
+      )
+    end
+  end
+
+  describe '#initialize' do
+    it 'activates a default LogRecordLimits' do
+      assert_equal(
+        OpenTelemetry::SDK::Logs::LogRecordLimits::DEFAULT,
+        logger_provider.instance_variable_get(:@log_record_limits)
       )
     end
   end
@@ -73,15 +82,15 @@ describe OpenTelemetry::SDK::Logs::LoggerProvider do
       # :version is nil by default, but explicitly setting it here
       # to make the test easier to read
       logger = logger_provider.logger(name: 'name', version: nil)
-      assert_equal(logger.instance_variable_get(:@instrumentation_scope).version, '')
+      assert_equal('', logger.instance_variable_get(:@instrumentation_scope).version)
     end
 
     it 'creates a new logger with the passed-in name and version' do
       name = 'name'
       version = 'version'
       logger = logger_provider.logger(name: name, version: version)
-      assert_equal(logger.instance_variable_get(:@instrumentation_scope).name, name)
-      assert_equal(logger.instance_variable_get(:@instrumentation_scope).version, version)
+      assert_equal(name, logger.instance_variable_get(:@instrumentation_scope).name)
+      assert_equal(version, logger.instance_variable_get(:@instrumentation_scope).version)
     end
   end
 
@@ -167,6 +176,51 @@ describe OpenTelemetry::SDK::Logs::LoggerProvider do
       OpenTelemetry::Common::Utilities.stub :maybe_timeout, 0 do
         logger_provider.add_log_record_processor(mock_log_record_processor)
         assert_equal(OpenTelemetry::SDK::Logs::Export::TIMEOUT, logger_provider.force_flush)
+      end
+    end
+  end
+
+  describe '#on_emit' do
+    let(:mock_context) do
+      mock_context = Minitest::Mock.new
+      def mock_context.value(key); OpenTelemetry::Trace::Span::INVALID; end # rubocop:disable Style/SingleLineMethods
+
+      mock_context
+    end
+
+    let(:args) do
+      span_context = OpenTelemetry::Trace::SpanContext.new
+      {
+        timestamp: Time.now,
+        observed_timestamp: Time.now + 1,
+        severity_text: 'DEBUG',
+        severity_number: 0,
+        body: 'body',
+        attributes: { 'a' => 'b' },
+        trace_id: span_context.trace_id,
+        span_id: span_context.span_id,
+        trace_flags: span_context.trace_flags,
+        instrumentation_scope: OpenTelemetry::SDK::InstrumentationScope,
+        context: mock_context
+      }
+    end
+
+    it 'creates a new log record' do
+      output = 'union station'
+      OpenTelemetry::SDK::Logs::LogRecord.stub(:new, ->(_) { puts output }) do
+        assert_output(/#{output}/) { logger_provider.on_emit(**args) }
+      end
+    end
+
+    it 'sends the log record to the processors' do
+      mock_log_record = Minitest::Mock.new
+
+      OpenTelemetry::SDK::Logs::LogRecord.stub :new, mock_log_record do
+        logger_provider.add_log_record_processor(mock_log_record_processor)
+        mock_log_record_processor.expect(:on_emit, nil, [mock_log_record, mock_context])
+
+        logger_provider.on_emit(**args)
+        mock_log_record_processor.verify
       end
     end
   end