+++ metrics_sdk/lib/opentelemetry/sdk/metrics/export/in_memory_metric_pull_exporter.rb
@@ -23,9 +23,9 @@ module OpenTelemetry
             export(collect)
           end
 
-          def export(metrics)
+          def export(metrics, timeout: nil)
             @mutex.synchronize do
-              @metric_snapshots << metrics
+              @metric_snapshots.concat(Array(metrics))
             end
             SUCCESS
           end