+++ metrics_sdk/lib/opentelemetry/sdk/metrics/meter.rb
@@ -11,6 +11,8 @@ module OpenTelemetry
     module Metrics
       # {Meter} is the SDK implementation of {OpenTelemetry::Metrics::Meter}.
       class Meter < OpenTelemetry::Metrics::Meter
+        NAME_REGEX = /\A[a-zA-Z][-.\w]{0,62}\z/
+
         # @api private
         #
         # Returns a new {Meter} instance.
@@ -52,12 +54,18 @@ module OpenTelemetry
 
         # @api private
         def add_metric_reader(metric_reader)
-          @instrument_registry.each do |_n, instrument|
+          @instrument_registry.each_value do |instrument|
             instrument.register_with_new_metric_store(metric_reader.metric_store)
           end
         end
 
         def create_instrument(kind, name, unit, description, callback)
+          raise InstrumentNameError if name.nil?
+          raise InstrumentNameError if name.empty?
+          raise InstrumentNameError unless NAME_REGEX.match?(name)
+          raise InstrumentUnitError if unit && (!unit.ascii_only? || unit.size > 63)
+          raise InstrumentDescriptionError if description && (description.size > 1023 || !utf8mb3_encoding?(description.dup))
+
           super do
             case kind
             when :counter then OpenTelemetry::SDK::Metrics::Instrument::Counter.new(name, unit, description, @instrumentation_scope, @meter_provider)
@@ -68,6 +76,11 @@ module OpenTelemetry
             when :observable_up_down_counter then OpenTelemetry::SDK::Metrics::Instrument::ObservableUpDownCounter.new(name, unit, description, callback, @instrumentation_scope, @meter_provider)
             end
           end
+        end
+
+        def utf8mb3_encoding?(string)
+          string.force_encoding('UTF-8').valid_encoding? &&
+            string.each_char { |c| return false if c.bytesize >= 4 }
         end
       end
     end