+++ metrics_sdk/test/opentelemetry/sdk/metrics/instrument/observable_up_down_counter_test.rb
@@ -21,7 +21,7 @@ describe OpenTelemetry::SDK::Metrics::Instrument::ObservableUpDownCounter do
     meter.create_observable_up_down_counter('updown_counter', unit: 'smidgen', description: 'a small amount of something', callback: callback)
 
     metric_exporter.pull
-    last_snapshot = metric_exporter.metric_snapshots.last
+    last_snapshot = metric_exporter.metric_snapshots
 
     _(last_snapshot[0].name).must_equal('updown_counter')
     _(last_snapshot[0].unit).must_equal('smidgen')
@@ -38,7 +38,7 @@ describe OpenTelemetry::SDK::Metrics::Instrument::ObservableUpDownCounter do
     up_down_counter.observe(timeout: 10, attributes: { 'foo' => 'bar' })
 
     metric_exporter.pull
-    last_snapshot = metric_exporter.metric_snapshots.last
+    last_snapshot = metric_exporter.metric_snapshots
 
     _(last_snapshot[0].name).must_equal('updown_counter')
     _(last_snapshot[0].unit).must_equal('smidgen')