+++ metrics_sdk/test/opentelemetry/sdk/metrics/meter_test.rb
@@ -76,7 +76,7 @@ describe OpenTelemetry::SDK::Metrics::Meter do
         _(counter_second.instance_variable_get(:@callbacks).size).must_equal 2
 
         metric_exporter.pull
-        last_snapshot = metric_exporter.metric_snapshots.last
+        last_snapshot = metric_exporter.metric_snapshots
 
         _(last_snapshot[0].name).must_equal('counter_first')
         _(last_snapshot[0].unit).must_equal('smidgen')
@@ -107,7 +107,7 @@ describe OpenTelemetry::SDK::Metrics::Meter do
         _(counter_second.instance_variable_get(:@callbacks).size).must_equal 2
 
         metric_exporter.pull
-        last_snapshot = metric_exporter.metric_snapshots.last
+        last_snapshot = metric_exporter.metric_snapshots
 
         _(last_snapshot[0].name).must_equal('counter_first')
         _(last_snapshot[0].unit).must_equal('smidgen')
@@ -128,8 +128,10 @@ describe OpenTelemetry::SDK::Metrics::Meter do
         # unregister the callback_second from instruments counter_first and counter_second
         meter.unregister([counter_first, counter_second], callback_second)
 
+
+        metric_exporter.reset
         metric_exporter.pull
-        last_snapshot = metric_exporter.metric_snapshots.last
+        last_snapshot = metric_exporter.metric_snapshots
 
         _(last_snapshot[0].name).must_equal('counter_first')
         _(last_snapshot[0].unit).must_equal('smidgen')
@@ -147,6 +149,67 @@ describe OpenTelemetry::SDK::Metrics::Meter do
         _(last_snapshot[1].data_points[0].attributes).must_equal({})
         _(last_snapshot[1].aggregation_temporality).must_equal(:delta)
       end
+    end
+  end
+
+  describe 'creating an instrument' do
+    INSTRUMENT_NAME_ERROR = OpenTelemetry::Metrics::Meter::InstrumentNameError
+    INSTRUMENT_UNIT_ERROR = OpenTelemetry::Metrics::Meter::InstrumentUnitError
+    INSTRUMENT_DESCRIPTION_ERROR = OpenTelemetry::Metrics::Meter::InstrumentDescriptionError
+    DUPLICATE_INSTRUMENT_ERROR = OpenTelemetry::Metrics::Meter::DuplicateInstrumentError
+
+    it 'duplicate instrument registration logs a warning' do
+      OpenTelemetry::TestHelpers.with_test_logger do |log_stream|
+        meter.create_counter('a_counter')
+        meter.create_counter('a_counter')
+        _(log_stream.string).must_match(/duplicate instrument registration occurred for instrument a_counter/)
+      end
+    end
+
+    it 'instrument name must not be nil' do
+      _(-> { meter.create_counter(nil) }).must_raise(INSTRUMENT_NAME_ERROR)
+    end
+
+    it 'instument name must not be an empty string' do
+      _(-> { meter.create_counter('') }).must_raise(INSTRUMENT_NAME_ERROR)
+    end
+
+    it 'instrument name must have an alphabetic first character' do
+      _(meter.create_counter('one_counter'))
+      _(-> { meter.create_counter('1_counter') }).must_raise(INSTRUMENT_NAME_ERROR)
+    end
+
+    it 'instrument name must not exceed 63 character limit' do
+      long_name = 'a' * 63
+      meter.create_counter(long_name)
+      _(-> { meter.create_counter(long_name + 'a') }).must_raise(INSTRUMENT_NAME_ERROR)
+    end
+
+    it 'instrument name must belong to alphanumeric characters, _, ., and -' do
+      meter.create_counter('a_-..-_a')
+      _(-> { meter.create_counter('a@') }).must_raise(INSTRUMENT_NAME_ERROR)
+      _(-> { meter.create_counter('a!') }).must_raise(INSTRUMENT_NAME_ERROR)
+    end
+
+    it 'instrument unit must be ASCII' do
+      _(-> { meter.create_counter('a_counter', unit: 'á') }).must_raise(INSTRUMENT_UNIT_ERROR)
+    end
+
+    it 'instrument unit must not exceed 63 characters' do
+      long_unit = 'a' * 63
+      meter.create_counter('a_counter', unit: long_unit)
+      _(-> { meter.create_counter('b_counter', unit: long_unit + 'a') }).must_raise(INSTRUMENT_UNIT_ERROR)
+    end
+
+    it 'instrument description must be utf8mb3' do
+      _(-> { meter.create_counter('a_counter', description: '💩'.dup) }).must_raise(INSTRUMENT_DESCRIPTION_ERROR)
+      _(-> { meter.create_counter('b_counter', description: "\xc2".dup) }).must_raise(INSTRUMENT_DESCRIPTION_ERROR)
+    end
+
+    it 'instrument description must not exceed 1023 characters' do
+      long_description = 'a' * 1023
+      meter.create_counter('a_counter', description: long_description)
+      _(-> { meter.create_counter('b_counter', description: long_description + 'a') }).must_raise(INSTRUMENT_DESCRIPTION_ERROR)
     end
   end
 end